Rename sysinfo module to system (#8356)
