--- x-pack/auditbeat/module/sysinfo/_meta/fields.yml
@@ -1,4 +1,0 @@
-- key: sysinfo
-  title: Sysinfo
-  description: These are the fields generated by the sysinfo module.
-  fields:
