+++ x-pack/auditbeat/module/system/_meta/config.yml.tmpl
@@ -1,6 +1,6 @@
 {{ if .Reference -}}
 {{ end -}}
-- module: sysinfo
+- module: system
   {{ if eq .GOOS "darwin" -}}
   metricsets:
     - host
