+++ x-pack/auditbeat/module/system/_meta/fields.yml
@@ -1,0 +1,4 @@
+- key: system
+  title: System
+  description: These are the fields generated by the system module.
+  fields:
